% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_addons.R
\name{add_individual_plant_info}
\alias{add_individual_plant_info}
\title{Add info on individual sites}
\usage{
add_individual_plant_info(df)
}
\arguments{
\item{df}{}
}
\value{
data.frame
}
\description{
Given a data.frame with columns PLANT_ID, this joins the following:
SITE_CODE, SPP_CODE, FUNC_GRP_CODE
}
\details{
Meant to be used internally in phenometR
}
