% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{add_year_doy_columns}
\alias{add_year_doy_columns}
\title{Add columns for the year and day of year (DOY)}
\usage{
add_year_doy_columns(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame
}
\description{
Given a data.frame with column DATE this adds two
columns YEAR and DOY
}
\details{
Meant to be used internally in phenometR
}
