% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{clear_credentials}
\alias{clear_credentials}
\title{Clear the stored database credentials}
\usage{
clear_credentials()
}
\value{
None
}
\description{
Clear the saved login information stored on your local machine. Try
    using this if your having trouble connecting to the database.
}
