% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{db_connect}
\alias{db_connect}
\title{Connect to the database}
\usage{
db_connect()
}
\value{
DBI Connecion object
}
\description{
Will prompt for credentials via get_credentials() if they are
not already stored.
}
