% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{get_credentials}
\alias{get_credentials}
\title{Get database credentials}
\usage{
get_credentials()
}
\value{
list of credential info
}
\description{
This will prompt for credentials first and store them, so
it only needs to ask for credentials once per session.
}
