% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{get_fg_phenophase}
\alias{get_fg_phenophase}
\title{Get all plant phenophases for a functional group.}
\usage{
get_fg_phenophase(
  functional_group,
  years = NULL,
  start_date = NULL,
  end_date = NULL,
  shape = "long"
)
}
\arguments{
\item{years}{Optional. integer or vector of integer for the years desired. years must be consecutive.}

\item{start_date}{Optional. A string with format 'YYYY-MM-DD'. Get visit information from this date forward. Default is all prior dates.}

\item{end_date}{Optional. A string with format 'YYYY-MM-DD'. Get visit information up to this date. Default is all dates up to todays date.}

\item{shape}{string. 'wide' or 'long' for a data.frame in the respective format. default 'long'}

\item{site_code}{string 2 letter site code}
}
\value{
data.frame
}
\description{
This retrieves all phenophase data for a specific functional group at all sites. This can be subset to specific years,
    or specific start and end dates.
    
    Functional Groups are:  
    'PG' - perennial grass
    'DS' - deciduous shrubs
    'ES' - evergreen shrubs
    'SU' - succulents
    
     If shape = 'long' (the default) then columns will be:   
     c('PLANT_ID','DATE','PHENOPHASE','STATUS','NOTE_FLAG','PHOTO_FLAG','SITE_CODE', 'SPP_CODE', 'FUNC_GRP_CODE')  
     
     If shape = 'wide' then all thephenophase codes will be columns.
}
\examples{
get_fg_phenophase(site_code = 'NO')
}
