% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenophase_metadata.R
\name{get_phenophase_metadata}
\alias{get_phenophase_metadata}
\title{Get phenophase metadata}
\usage{
get_phenophase_metadata()
}
\value{
data.frame of phenophase metadata
}
\description{
Get a data.frame describing all the phenophases codes.

The phenophase metadata includes the columns:
functional_group, phenophase_type, phenophase, phenophase_desc
}
\examples{
get_phenophase_metadata()
}
