% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{get_plant_info}
\alias{get_plant_info}
\title{Get all individual plant info}
\usage{
get_plant_info()
}
\value{
data.frame of plant_id, site_code, species, functional group, etc.
}
\description{
Retrieve information on all indvidual plants.
}
