% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{get_site_list}
\alias{get_site_list}
\title{Get site information}
\usage{
get_site_list()
}
\value{
data.frame of site information
}
\description{
Retrieve information on all the sites.
}
\examples{
get_site_list()
}
