% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{get_site_phenophase}
\alias{get_site_phenophase}
\title{Get all plant phenophases for a site.}
\usage{
get_site_phenophase(
  site_code,
  years = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{site_code}{string. 2 letter site code}

\item{years}{Optional. integer or vector of integer for the years desired. years must be consecutive.}

\item{start_date}{Optional. A string with format 'YYYY-MM-DD'. Get visit information from this date forward. Default is all prior dates.}

\item{end_date}{Optional. A string with format 'YYYY-MM-DD'. Get visit information up to this date. Default is all dates up to todays date.}
}
\value{
data.frame
}
\description{
Retrieve phenophase data for all plants at a specific site. Can be subset to specific years,
    or specific start and end dates. Since sites include multiple functional groups, 
    this only returns in the long format with columns:
    
     c('PLANT_ID','DATE','PHENOPHASE','STATUS','NOTE_FLAG','PHOTO_FLAG','SITE_CODE', 'SPP_CODE', 'FUNC_GRP_CODE')
}
\examples{
get_site_phenophase(site_code = 'NO')
}
