% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{get_site_visits}
\alias{get_site_visits}
\title{Get site visit information}
\usage{
get_site_visits(site_codes = NULL, start_date = NULL, end_date = NULL)
}
\arguments{
\item{site_codes}{Optional. A single site code or list of site codes. Default returns all sites.}

\item{start_date}{Optional. A string with format 'YYYY-MM-DD'. Get visit information from this date forward. Default is all prior dates.}

\item{end_date}{Optional. A string with format 'YYYY-MM-DD'. Get visit information up to this date. Default is all dates up to todays date.}
}
\value{
data.frame of site visit information.
}
\description{
Get site visit information
}
