% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{get_species_phenophase}
\alias{get_species_phenophase}
\title{Get phenophase information for a single species.}
\usage{
get_species_phenophase(
  spp_code,
  years = NULL,
  start_date = NULL,
  end_date = NULL,
  shape = "long"
)
}
\arguments{
\item{spp_code}{string. unique species identifier}

\item{years}{Optional. integer or vector of integer for the years desired. years must be consecutive.}

\item{start_date}{Optional. A string with format 'YYYY-MM-DD'. Get visit information from this date forward. Default is all prior dates.}

\item{end_date}{Optional. A string with format 'YYYY-MM-DD'. Get visit information up to this date. Default is all dates up to todays date.}

\item{shape}{string. 'wide' or 'long' for a data.frame in the respective format. default 'long'}
}
\value{
data.frame of phenophases by date
}
\description{
This function retrieves all phenophase data for a single species. All individual at all sites for the
     species will be return. Can be subset to specific years, or specific start and end dates. 
     
     See the output from get_plant_info() for valid species codes. 
    
     If shape = 'long' (the default) then columns will be:   
     c('PLANT_ID','DATE','PHENOPHASE','STATUS','NOTE_FLAG','PHOTO_FLAG','SITE_CODE', 'SPP_CODE', 'FUNC_GRP_CODE')  
     
     If shape = 'wide' then all thephenophase codes will be columns.
}
\examples{
get_species_phenophase('PRGL')
get_species_phenophase('PRGL', start_date = '2012-01-01', end_date = '2012-12-31')
}
