% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{is_error}
\alias{is_error}
\title{Helper function}
\usage{
is_error(x)
}
\value{
bool
}
\description{
Use with try to detect and handle errors
}
\details{
con = try(DBIConnect(..))
if(is_error(con)){
  handle error...
}
}
