% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{parse_dates}
\alias{parse_dates}
\title{Parse date options}
\usage{
parse_dates(years, start_date, end_date)
}
\arguments{
\item{years}{int or vector of ints for the years desired. years must be consecutive.}

\item{start_date}{str in the form 'YYYY-MM-DD'}

\item{end_date}{str in the form 'YYYY-MM-DD'}
}
\value{
list with start_date end_date options.
}
\description{
Used internally to parse the date arguments and return a start/end date to use in the db filter.
}
\details{
If years is set then create a start/end date for the db query. In this case start_date/end_date args will be ignored.
If all is NULL then set end_date to the current date and start_date at the earliest possible date.
}
