% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{set_phenometR_backend}
\alias{set_phenometR_backend}
\title{Set the database backend.}
\usage{
set_phenometR_backend(backend)
}
\arguments{
\item{backend}{string. one of 'jornada-server' or 'test-db'}
}
\description{
Set by default to the primary jornada server. 
Use this to set to a local sqlite db
for testing only
}
