% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{get_phenophase_metadata}
\alias{get_phenophase_metadata}
\title{Get phenophase metadata}
\usage{
get_phenophase_metadata(full_metadata = FALSE)
}
\arguments{
\item{functional_groups}{boolean. If TRUE return all metadata info, if FALSE (default) return only attribute name and definition.}
}
\value{
data.frame of phenophase metadata
}
\description{
Get a data.frame describing all the phenophases codes.
    eg. for the phenophase code GS_01 obtain the description 
    "Initial growth for perennial grasses", among other things.

The full metadata info includes:
ATTRIBUTE_ID,ATTRIBUTE_NAME,ATTRIBUTE_DEFINITION,ATTRIBUTE_DATA_TYPE,NULL_VALUE,DESCRIPTION
}
\examples{
get_phenophase_metadata()
}
